support mocks addon and decorator in Storybook 7
