+++ .changeset/lazy-bags-accept.md
@@ -1,0 +1,5 @@
+---
+'@web/mocks': patch
+---
+
+support Storybook 7
