+++ .changeset/thin-actors-clean.md
@@ -1,0 +1,5 @@
+---
+'@web/storybook-utils': major
+---
+
+create a package for Storybook utilities, move the utility to help make Storybook addons with web components from @web/storybook-prebuilt to reuse in Storybook 7
