+++ packages/storybook-utils/src/index.js
@@ -1,0 +1,1 @@
+export { createAddon } from './create-addon.js';
