+++ packages/storybook-utils/tsconfig.json
@@ -1,0 +1,24 @@
+// Don't edit this file directly. It is generated by generate-ts-configs script
+
+{
+  "extends": "../../tsconfig.browser-base.json",
+  "compilerOptions": {
+    "module": "ESNext",
+    "outDir": "./dist",
+    "rootDir": "./src",
+    "composite": true,
+    "allowJs": true,
+    "checkJs": true,
+    "emitDeclarationOnly": true
+  },
+  "references": [],
+  "include": [
+    "src",
+    "types"
+  ],
+  "exclude": [
+    "src/browser",
+    "tests",
+    "dist"
+  ]
+}
