+++ workspace-packages.mjs
@@ -17,6 +17,7 @@ const packages = [
   { name: 'dev-server-import-maps', type: 'ts', environment: 'node' },
   { name: 'storybook-builder', type: 'ts', environment: 'node' },
   { name: 'storybook-framework-web-components', type: 'ts', environment: 'node' },
+  { name: 'storybook-utils', type: 'js', environment: 'browser' },
   { name: 'test-runner', type: 'ts', environment: 'node' },
   { name: 'test-runner-core', type: 'ts', environment: 'node' },
   { name: 'test-runner-chrome', type: 'ts', environment: 'node' },
